Integrado, mas sem banco novo

+++ Repositories/PlantioRepository.cs
@@ -1,49 +1,73 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Plantech.Data;
 using Plantech.Interfaces;
 using Plantech.Models;
 
-namespace Plantech.Repositories
-{
-    public class PlantioRepository(PlantechContext context) : IPlantioRepository
+public class PlantioRepository(PlantechContext context) : IPlantioRepository
 {
     private readonly PlantechContext _context = context;
-
-        public async Task<IEnumerable<Plantio>> GetAllAsync()
+    
+    public async Task<Plantio> GetUltimoPlantioAsync()
     {
-        return await _context.Plantios.ToListAsync();
+        return await _context.Plantios
+            .OrderByDescending(p => p.Id)
+            .FirstOrDefaultAsync();
     }
+
+    public async Task<IEnumerable<Plantio>> GetAllAsync()
+    {
+        return await _context.Plantios.Include(p => p.Hortalica).ToListAsync();
+    }
+
 
     public async Task<Plantio> GetByIdAsync(int id)
     {
         return await _context.Plantios.FindAsync(id);
     }
 
-    public async Task AddAsync(Plantio plantio)
-    {
-        await _context.Plantios.AddAsync(plantio);
-        await _context.SaveChangesAsync();
-    }
-
     public async Task UpdateAsync(Plantio plantio)
     {
-        _context.Plantios.Update(plantio);
-        await _context.SaveChangesAsync();
+        if (PlantioExists(plantio.Id))
+        {
+            _context.Plantios.Update(plantio);
+            await _context.SaveChangesAsync();
+        }
     }
 
     public async Task DeleteAsync(int id)
     {
-        var plantio = await _context.Plantios.FindAsync(id);
-        if (plantio != null)
+        if (PlantioExists(id))
         {
-            _context.Plantios.Remove(plantio);
-            await _context.SaveChangesAsync();
+            var plantio = await _context.Plantios.FindAsync(id);
+            if (plantio != null)
+            {
+                _context.Plantios.Remove(plantio);
+                await _context.SaveChangesAsync();
+            }
         }
     }
-}
+
+    public async Task CreatePlantioAsync(Plantio plantio)
+    {
+        await _context.Plantios.AddAsync(plantio);
+        await _context.SaveChangesAsync(); 
+    }
+
+    public async Task CreateInsumosPlantioAsync(InsumosPlantio insumoPlantio)
+    {
+        await _context.InsumosPlantios.AddAsync(insumoPlantio);
+        await _context.SaveChangesAsync(); 
+    }
+
+    private bool PlantioExists(int id)
+    {
+        return _context.Plantios.Any(e => e.Id == id);
+    }
+
+    // Lotes de insumo
+    public async Task<IEnumerable<LotesInsumo>> GetLotesInsumosAsync()
+    {
+        return await _context.LotesInsumos.ToListAsync();
+    }
 
 }
